feat: Add camelCaseToWords method to Executor class
The code changes add a new method, camelCaseToWords, to the Executor class. This method converts a camelCase string to words with spaces between each word. It is used to format the name of a patient expression when processing DocumentReference sources. This commit also includes updates to the cql4browsers.js and executor.ts files.